ARM: dts: add rk3288-firefly-reload
The Firefly RK3288 Reload is a combination Firefly rk3288 core board
with the Reload baseboard. Add a dtsi for the Firefly rk3288 core
which can be included into the dts for the various baseboards
in the future and dts for Reload base board.

Currently supported are serial console, wired networking, eMMC and
SD storage, SPFIF, IR receiver, LEDs, SDIO wifi and USB. But only
the OTG could work on the host mode now, the other USB host can't
work now, additional patches are required.

Signed-off-by: Randy Li <ayaka@soulik.info>