ARM: dts: rockchip: add the video codec nodes at RK3288
The VPU IP is a video encoder/decoder acceleration unit
existing at RK3288. And the HEVC found on rk3288 platform
is used for HEVC decoding.

Please notice there is a different in configure memory range
between the one used in kernel 4.4.

In order to solve the complex video IP combo and configure
problem, I decide to make each features independent.

Signed-off-by: Randy Li <randy.li@rock-chips.com>
Signed-off-by: Randy Li <ayaka@soulik.info>