video: rockchip/mpp_device: add VPU decoder
There are two modules for the different generation of
the VPU decoder. The register of  the VPU decoder
usually have an offset to the VPU base.

Please notice that the encoder of the VPU is supported
in the other module.

The decoder and encoder of the VPU usually share some
hardware resources(Not the RK3328, the encoders at that
platform share the resources), so they will be regarded
as an combo sharing the same mpp service in this design
of the mpp driver.

The Post Processor is not supported now (Only work for
 JPEG decoder). I am planning to move it into the
other module.

Change-Id: Ic5d1f4f6ad9b58bc805e6edeace3a5fa23d74e20
Signed-off-by: ayaka <ayaka@soulik.info>
Signed-off-by: Randy Li <randy.li@rock-chips.com>