video: rockchip/mpp_service: switching in multiple sequences
--------------    ---------------    ------------------
|            |    |             |    | mpp device     |
| mpp device |<-->| mpp service |<-->|  implementation|
|    top     |    |             |    |  and buttom    |
--------------    ---------------    ------------------

The device class of the mpp service will be create at
the module loading time. But the class for the mpp device
will create at the mpp common device.

The power and clocks management have been left to the
power domain, so the combo device won't worry whether
those things are ready when the device is going to run.

It won't work for the link table mode. The steps of the lock
are fine at the most of time, but we need a way to push
a bunch of tasks to run then wait them done. I also need a way
to insert new task.

TODO:
1. Device scheduler

Change-Id: I4abae5fa054a583fbb4a49fc68726ce8405f2901
Signed-off-by: Randy Li <randy.li@rock-chips.com>
Signed-off-by: ayaka <ayaka@soulik.info>